refactor(PR4-A): converge state file naming to state_X.go prefix; fix examplechild/exampleslow package names
