+++ umh-core/pkg/fsmv2/workers/example/exampleslow/worker.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package example_slow
+package exampleslow
 
 import (
 	"context"
