+++ umh-core/pkg/fsmv2/workers/example/helloworld/state/state_degraded.go
