+++ umh-core/pkg/fsmv2/workers/persistence/state/state_stopped.go
