+++ umh-core/pkg/fsmv2/workers/transport/state/state_degraded.go
